调整了采样方法以减小锯齿, 改变了相机的参数

+++ Path-Tracing/vector.h
@@ -52,4 +52,4 @@ public:
 	}
 };
 
-typedef Vec3<float> Vec3f;
+typedef Vec3<double> Vec3f;
